pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-builder"
version = "0.1.0"
description = "CLI tool to analyze repositories and generate MCP YAML files"
authors = [{name = "Principal Engineer", email = "engineer@example.com"}]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
local-llm = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
]
all = [
    "openai>=1.0.0", 
    "anthropic>=0.18.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
]
requires-python = ">=3.8"

[project.scripts]
mcp-builder = "mcp_builder.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_builder*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]